% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getBeta}
\alias{getBeta}
\alias{getBetaMetrics}
\title{Beta}
\usage{
getBeta(x, id)

getBetaMetrics(x, ab)
}
\arguments{
\item{x}{(data.frame) Has to have columns Species, YEAR, assemblageID,
STUDY_ID, cell and Abundance or Biomass}

\item{id}{definition of id}

\item{ab}{character input for chosen currency - "A" = Abundance or "B" = Biomass}
}
\value{
getBeta returns a data.frame with three beta diversity dissimilarity
metrics

getBetaMetrics returns a long data.frame with results for three beta
  metrics.
}
\description{
Beta

run the beta function
}
\examples{
\dontrun{
  x <- data.frame(
    YEAR = rep(rep(2010:2015, each = 4), times = 4),
    matrix(data = rpois(384, 2), ncol = 4)
  )
  res <- getBeta(x, "A")
}
\dontrun{
x <- data.frame(
  YEAR = rep(rep(2010:2015, each = 4), times = 4),
  Species = c(replicate(
   n = 8L,
   sample(letters, 24L, replace = FALSE))),
  Abundance = rpois(24 * 8, 10),
  assemblageID = rep(LETTERS[1L:8L], each = 24)
  )

res <- getBetaMetrics(x, "A")
}
}
\author{
Faye Moyes
}
