% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcolorsBioTIME.R
\name{scale_color_biotime}
\alias{scale_color_biotime}
\alias{scale_colour_biotime}
\alias{scale_fill_biotime}
\title{Scale construction for ggplot use}
\usage{
scale_color_biotime(palette = "realms", discrete = TRUE, reverse = FALSE, ...)

scale_colour_biotime(palette = "realms", discrete = TRUE, reverse = FALSE, ...)

scale_fill_biotime(palette = "realms", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{One of: `realms`, `gradient`, `cool`, `warm`, default to `realms`.}

\item{discrete}{See Details. default to `FALSE`}

\item{reverse}{Default to `FALSE`}

\item{...}{Passed to \code{\link[ggplot2]{discrete_scale}} or \code{\link[ggplot2]{scale_color_gradient}}}
}
\value{
If \code{discrete} is \code{TRUE}, the function returns a colour palette produced by 
    \code{\link[ggplot2]{discrete_scale}} and if \code{discrete} is \code{FALSE}, the function
    returns a colour palette produced by \code{\link[ggplot2]{scale_color_gradient}}.

If \code{discrete} is \code{TRUE}, the function returns a colour palette produced by 
    \code{\link[ggplot2]{discrete_scale}} and if \code{discrete} is \code{FALSE}, the function
    returns a colour palette produced by \code{\link[ggplot2]{scale_color_gradient}}.
}
\description{
Scale construction for ggplot use

Scale construction for filling in ggplot
}
\details{
USAGE NOTE: Remember to change these arguments when plotting colours continuously.
}
\author{
Cher F. Y. Chow
}
