% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsBioTIME.R
\name{plotSlopes}
\alias{plotSlopes}
\alias{themeBioTIME}
\title{Plot slopes BioTIME}
\usage{
plotSlopes(x, metric, cols, taxa, method, assemblageID, divType)

themeBioTIME(lp, fontSize, colx, coly)
}
\arguments{
\item{x}{Parameter description}

\item{metric}{Parameter description}

\item{cols}{Parameter description}

\item{taxa}{Parameter description}

\item{method}{Parameter description}

\item{assemblageID}{Parameter description}

\item{divType}{Parameter description}

\item{lp}{description}

\item{fontSize}{description}

\item{colx}{description}

\item{coly}{description}
}
\value{
A plot
}
\description{
Plot slopes BioTIME

ggplot2 theme for BioTIME plots
}
\examples{
\dontrun{
library(dplyr)
library(tidyr)
library(ggplot2)

load("./data/subBTmeta.RData")
meta  <-  subBTmeta
meta2 <- select(meta, STUDY_ID, TAXA, REALM)

alphaSl <- as.data.frame(alphaSl \%>\% separate(., assemblageID,
                                            into= c("STUDY_ID", "cell"), sep="_", remove=F))

ameta <- merge(alphaSl, meta2, by="STUDY_ID")

betaSl <- as.data.frame(betaSl \%>\% separate(., assemblageID,
                                          into= c("STUDY_ID", "cell"), sep="_", remove=F))

bmeta <- merge(betaSl, meta2, by="STUDY_ID")
bio <- c("#00483d","#127c8e","#31b9c2","#86db9c","#d9d956" ,"#c0f176","#ffff67", "#cf7941")
bioCol <- c("#cf7941", "#127c8e","#86db9c","#d9d956")

plm <- plotSlopes(bmeta, "JaccardDiss", bioCol, "Fish", "metric", "18_9044316", "beta")

plt <- plotSlopes(ameta, "N", bio, "Fish", "taxa","18_9044316", "alpha")

plrf <- plotSlopes(bmeta, "N", bioCol, "Fish", "ind","18_9044316", "beta")
}
\dontrun{
fig1 <- ggplot2::ggplot() +
          themeBioTIME("none", 12, "black", "grey90")
}
}
