% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BTsubset_data}
\alias{BTsubset_data}
\title{BioTIME subset}
\format{
## `BTsubset_data`
A data frame with 81,084 rows and 17 columns:
\describe{
  \item{ID_ALL_RAW_DATA}{Unique BioTIME identifier for record}
  \item{ABUNDANCE}{Double representing the abundance for the record (see metadata for details of ABUNDANCE_TYPE}
  \item{BIOMASS}{Double representing the biomass for the record (see metadata for details of BIOMASS_TYPE}
  \item{ID_SPECIES}{Unique identifier linking to the species table}
  \item{SAMPLE_DESC}{Concatenation of variables comprising unique sampling event}
  \item{PLOT}{Name or identifier of plot field, only used for fixed plots such as forest quadrats}
  \item{LATITUDE}{Latitude of record}
  \item{LONGITUDE}{Longitude of record}
  \item{DEPTH}{Depth or elevation of record if available}
  \item{DAY}{Numerical day of record}
  \item{MONTH}{Numerical value of month for record, i.e. January=1}
  \item{YEAR}{Year of record}
  \item{STUDY_ID}{BioTIME study unique identifier}
  \item{newID}{Validated species identifier key}
  \item{valid_name}{Highest taxonomic resolution of individual, preferred is genus and species}
  \item{resolution}{Level of resolution, i.e. 'species' represented by genus and species}
  \item{taxon}{Higher level taxonomic grouping, i.e. Fish}
}
}
\source{
<https://biotime.st-andrews.ac.uk/download.php>
}
\usage{
BTsubset_data
}
\description{
A subset of data from BioTIME temporal surveys.
}
\keyword{datasets}
