% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{doPivot}
\alias{doPivot}
\title{Convert table long to wide}
\usage{
doPivot(x)
}
\arguments{
\item{x}{(data.frame) Has to have columns Species and Abundance}
}
\value{
Wide x with species in columns.
}
\description{
Convert table long to wide
}
\author{
Faye Moyes
}
