% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getAlpha}
\alias{getAlpha}
\title{Alpha}
\usage{
getAlpha(x)
}
\arguments{
\item{x}{(\code{data.frame}) First column has to be year and following columns
contain species abundances.}
}
\value{
A data frame with results for S (species richness), N (numerical abundance),
   maximum N per year per assemblage, Shannon, Exponential Shannon, Simpson,
   Inverse Simpson, PIE (probability of intraspecific encounter) and
   McNaughton's Dominance.
}
\description{
Alpha
}
\examples{
\dontrun{
# 1 site, 1 year in long format, ordered by ABUNDANCE or BIOMASS
x <- data.frame(species = letters[1:6], x = 6:1)
getAlpha(x$x)
}
}
\keyword{internal}
