% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getAlphaMetrics_reference}
\alias{getAlphaMetrics_reference}
\title{Alpha diversity metrics}
\usage{
getAlphaMetrics_reference(x, measure)
}
\arguments{
\item{x}{(\code{data.frame}) BioTIME data in the format of the output of the
\code{\link{resampling}} function.}

\item{measure}{(\code{character}) chosen currency defined by a single column name.}
}
\description{
Calculates a set of standard alpha diversity metrics
}
\examples{
  # Mean and sd values of the metrics for several resamplings
  gridding(BTsubset_meta, BTsubset_data) |>
    resampling(measure = "BIOMASS", resamps = 2) |>
    dplyr::reframe(
       getAlphaMetrics_reference(
          x = pick(assemblageID, YEAR, Species, BIOMASS),
          measure = "BIOMASS"),
       .by = resamp) |>
    dplyr::summarise(
       dplyr::across(
          .cols = !resamp,
          .fns = c(mean = mean, sd = sd)),
       .by = c(assemblageID, YEAR)) |>
    tidyr::pivot_longer(
       col = dplyr::contains("_"),
       names_to = c("metric", "stat"),
       names_sep = "_",
       names_transform = as.factor) |>
    tidyr::pivot_wider(names_from = stat) |>
    head(10)
}
