% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getBeta}
\alias{getBeta}
\title{Beta}
\usage{
getBeta(x, id)
}
\arguments{
\item{x}{(data.frame) First column has to contain year values and following
columns contain species abundances}

\item{id}{definition of id}
}
\value{
getBeta returns a data.frame with three beta diversity dissimilarity
metrics
}
\description{
Beta
}
\examples{
\dontrun{
  x <- data.frame(
    YEAR = rep(rep(2010:2015, each = 4), times = 4),
    matrix(data = rpois(384, 2), ncol = 4)
  )
  res <- getBeta(x, "A")
}
}
\author{
Faye Moyes
}
\keyword{internal}
