% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getBetaDissimilarity}
\alias{getBetaDissimilarity}
\title{run the beta function}
\usage{
getBetaDissimilarity(x)
}
\arguments{
\item{x}{(data.frame) Has to have columns Species and Abundance}
}
\value{
long with results for three beta metrics
}
\description{
run the beta function
}
\author{
Faye Moyes
}
