% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getBetaMetrics_reference}
\alias{getBetaMetrics_reference}
\title{Beta diversity metrics}
\usage{
getBetaMetrics_reference(x, measure)
}
\arguments{
\item{x}{(\code{data.frame}) BioTIME data table in the format of the output of the
\code{\link{resampling}} functions.}

\item{measure}{(\code{character}) chosen currency defined by a single column name.}
}
\description{
Calculates a set of standard beta diversity metrics
}
\examples{
gridding(BTsubset_meta, BTsubset_data) |>
  resampling(measure = "BIOMASS", verbose = FALSE, resamps = 2) |>
  dplyr::reframe(
       getBetaMetrics_reference(
          x = pick(assemblageID, YEAR, Species, BIOMASS),
          measure = "BIOMASS"),
       .by = resamp)

}
