% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopesBioTIME.R
\name{getLinearRegressions}
\alias{getLinearRegressions}
\title{Get Linear Regressions BioTIME}
\usage{
getLinearRegressions(x, divType)
}
\arguments{
\item{x}{Parameter description}

\item{divType}{Parameter description}
}
\value{
A plot
}
\description{
Get Linear Regressions BioTIME
}
