% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridBioTIME.R
\name{gridding}
\alias{gridding}
\title{gridding BioTIME
Grids BioTIME data into 96km^2^ cells based on location (latitude/longitude)}
\usage{
gridding(meta, btf, res = 12L)
}
\arguments{
\item{meta}{(data.frame) BioTIME metadata}

\item{btf}{(data.frame) BioTIME data}

\item{res}{(integer) XXXthe resolution for the grid sizeXXX}
}
\value{
Returns a data.frame, with selected columns from the btf and meta
  data.frames, an extra integer column called cell and two character columns
  called StudyMethod and AssemblageID (concatenation of study_ID and cell).
}
\description{
gridding BioTIME
Grids BioTIME data into 96km^2^ cells based on location (latitude/longitude)
}
\details{
Each BioTIME study contains distinct samples measured (with a
  consistent methodology) over time, which could be fixed plots (i.e. SL or
  "single-location" studies where measures are taken from a set of specific
  georeferenced sites at any given time) or wide-ranging surveys, transects,
  tows, and so on (i.e. ML or "multi-location" studies where measures are taken
  from multiple sites that may or may not align from year to year, see
  \link{runResampling}). `gridding()` is a function designed to identify,
  separate and standardize both SL and ML studies using a global grid of 96km2
  hexagonal cells derived from \code{\link[dggridR]{dgconstruct}} (res=12).
  Here each sample is assigned a different combination of study ID and grid cell
  (based on its latitude and longitude) resulting in a unique identifier for
  each assemblage time-series (AssemblageID), thus allowing for the integrity
  of each study and each sample to be maintained.
  By default `meta` represents a long form data frame containing the data
  information for BioTIME studies and `btf` is a data frame containing long
  form data from a main BioTIME query.
}
\examples{
\dontrun{
  library(BioTIMEr)
  gridding(subBTmeta, subBTquery)
}
}
