% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_BioTIME.R
\name{parser_BioTIME}
\alias{parser_BioTIME}
\title{Function splitting SAMPLE_DESC in various columns based on SAMPLE_DESC_NAME}
\usage{
<<<<<<< HEAD
parser_BioTIME(x, format, pattern, verbose = TRUE)
=======
parser_BioTIME(x, format, pattern)
>>>>>>> c44707e9b6912963553b5902c1cf86cec60ca30b
}
\arguments{
\item{x}{SAMPLE_DESC or a column that contains data values to split}

\item{format}{SAMPLE_DESC_NAME or a column containing the structure or format
to organise codes into columns.}

\item{pattern}{A string to split \code{x} and \code{format} by. Passed to
\code{\link[stringi]{stri_split_fixed}}.}
}
\value{
<<<<<<< HEAD
A \code{data.frame} with column names extracted from the
\code{format} strings. If there were mismatches between \code{x} and
\code{format} strings, \code{error_x} and \code{error_format} columns show
problematic values.
=======
A \code{data.frame}
>>>>>>> c44707e9b6912963553b5902c1cf86cec60ca30b
}
\description{
Function splitting SAMPLE_DESC in various columns based on SAMPLE_DESC_NAME
}
\examples{
   x <- c("T1_23.1212_56.2346_1995_12_30", "786_-60.4567_0.346_2027")
   format <- c("transect_lon_lat_year_month_day", "depth_lat_lon_year")
   parser_BioTIME(x, format, "_")

  dt <- data.frame(
    code = rep(
      c("T1_23.1212_56.2346_1995_12_30","786_-60.4567_0.346_2027"),
      each = 10
    ),
    name = rep(
      c("transect_lon_lat_year_month_day", "depth_lat_lon_year"),
      each = 10
    )
  )

  parser_BioTIME(dt$code, dt$name, "_")

  dplyr::mutate(dt, parser_BioTIME(code, name, "_"))

<<<<<<< HEAD
=======

>>>>>>> c44707e9b6912963553b5902c1cf86cec60ca30b
}
