% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBioTIME.R
\name{parseBioTIME}
\alias{parseBioTIME}
\alias{simplify_SAMPLE_DESC}
\title{Function splitting SAMPLE_DESC in various columns based on SAMPLE_DESC_NAME}
\usage{
parseBioTIME(x, format, sep, verbose = TRUE)

simplify_SAMPLE_DESC(x, format, sep, exclude)
}
\arguments{
\item{x}{SAMPLE_DESC or a vector of strings of the form
'T1_23.1212_56.2346_1995_12_30'}

\item{format}{SAMPLE_DESC_NAME or a vector of strings of the form
'depth_lat_lon_year'}

\item{sep}{\code{string}}

\item{verbose}{if FALSE, discrpeancies between \code{c} and \code{format}
will not be shown.}

\item{exclude}{\code{character} vector of elements to remove}
}
\value{
A \code{data.frame} with column names extracted from the
\code{format} strings. If there were mismatches between \code{x} and
\code{format} strings, \code{error_x} and \code{error_format} columns show
problematic values.

a vector of the same length as \code{x} with strings from \code{x}
   where values in \code{exclude} were removed.
}
\description{
Function splitting SAMPLE_DESC in various columns based on SAMPLE_DESC_NAME

Removing specific elements of SAMPLE_DESC
}
\examples{
   x <- c("T1_23.1212_56.2346_1995_12_30", "786_-60.4567_0.346_2027")
   format <- c("transect_lon_lat_year_month_day", "depth_lat_lon_year")
   parseBioTIME(x, format, "_")

  dt <- data.frame(
    code = rep(
      c("T1_23.1212_56.2346_1995_12_30","786_-60.4567_0.346_2027"),
      each = 10
    ),
    name = rep(
      c("transect_lon_lat_year_month_day", "depth_lat_lon_year"),
      each = 10
    )
  )

  parseBioTIME(dt$code, dt$name, "_")

  dplyr::mutate(dt, parseBioTIME(code, name, "_"))

x <- c("T1_23.1212_56.2346_1995_12_30", "786_-60.4567_0.346_2027")
format <- c("transect_lon_lat_year_month_day", "depth_lat_lon_year")
exclude <- c("YEAR", "year", "month", "day")
sep = "_"
simplify_SAMPLE_DESC(x, format, "_", "year")
}
