% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyBioTIME.R
\name{rarefysamples}
\alias{rarefysamples}
\title{rarefy BioTIME}
\usage{
rarefysamples(Year, SampleID, Species, Abundance, resamps)
}
\arguments{
\item{Year}{(character) string with latitude and longitude, one or more in a vector.}

\item{SampleID}{A definition of SampleID}

\item{Species}{A definition of Species}

\item{Abundance}{A definition of Abundance. All the same length?}

\item{resamps}{Number of repetition}
}
\value{
A data.frame with rarefied abundances
}
\description{
rarefy BioTIME
}
