% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyBioTIME.R
\name{resampling}
\alias{resampling}
\title{resampling BioTIME
Uses the output of `gridding` and applies the `rarefysamples` function.}
\usage{
resampling(df, measure, resamps = 1L, conservative = FALSE)
}
\arguments{
\item{df}{`data.frame` to be resampled (in the format of the output of the
\code{\link{gridding}} function).}

\item{measure}{(character) Enter the names of the columns of interest:
"ABUNDANCE", "BIOMASS" or c("ABUNDANCE","BIOMASS"). The column cannot have 0s.}

\item{resamps}{(integer) Number of repetitions, 1 by default.}

\item{conservative}{(logical) If `TRUE`, whenever a NA is found for abundance
or biomass, the whole sample is removed instead of the missing observations
only. `FALSE` by default.}
}
\value{
`data.frame` containing rarefied studies
}
\description{
resampling BioTIME
Uses the output of `gridding` and applies the `rarefysamples` function.
}
\examples{
\dontrun{
  library(BioTIMEr)
  set.seed(42)
  df <- gridding(subBTmeta, subBTquery)
  resampling(df, measure = "BIOMASS")
  resampling(df, measure = "ABUNDANCE")
  resampling(df, measure = c("ABUNDANCE","BIOMASS"))
}

}
