% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subBTquery}
\alias{subBTquery}
\title{BioTIME subset}
\format{
## `subBTquery`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
subBTquery
}
\description{
A subset of data from BioTIME temporal surveys.
}
\keyword{datasets}
