% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getAlpha}
\alias{getAlpha}
\alias{getAlphaMetrics}
\title{Alpha}
\usage{
getAlpha(x, id)

getAlphaMetrics(x, ab)
}
\arguments{
\item{x}{(data.frame) First column has to be year}

\item{id}{definition of id}

\item{ab}{character input for chosen currency - "A" = Abundance or "B" = Biomass}
}
\value{
A data frame with results for S (species richness), N (numerical abundance),
maximum N per year per assemblage, Shannon, Exponential Shannon, Simpson,
Inverse Simpson, PIE (probability of intraspecific encounter) and
McNaughton's Dominance.

getAlphaMetrics returns a data.frame with nine alpha diversity metrics
}
\description{
Alpha

run the alpha function
}
\examples{
\dontrun{
  x <- data.frame(
    Year = rep(rep(2010:2015, each = 4), times = 4),
    matrix(data = rpois(384, 10), ncol = 4)
    )
    res <- getAlpha(x, rarefyID = 34)
}

\dontrun{
  x <- data.frame(
    Year = rep(rep(2010:2015, each = 4), times = 4),
    Species = c(replicate(n = 8L, sample(letters, 24L, replace = FALSE))),
    Abundance = rpois(24 * 8, 10),
    rarefyID = rep(LETTERS[1L:8L], each = 24)
  )
  res <- getAlphaMetrics(x, "A")
}
}
\author{
Faye Moyes
}
