% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getBeta}
\alias{getBeta}
\alias{getBetaDissimilarity}
\title{Beta}
\usage{
getBeta(x, id)

getBetaDissimilarity(x, ab)
}
\arguments{
\item{x}{(data.frame) Has to have columns Species and Abundance}

\item{id}{definition of id}

\item{ab}{character input for chosen currency - "A" = Abundance or "B" = Biomass}
}
\value{
getBeta returns a data.frame with three beta diversity dissimilarity metrics

getBetaDissimilarity returns a long data.frame with results for three beta metrics
}
\description{
Beta

run the beta function
}
\examples{
\dontrun{
res <- getBetaDissimilarity(x, "A")
where x is a long form data frame
}
}
\author{
Faye Moyes
}
