% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getBeta}
\alias{getBeta}
\alias{getBetaDissimilarity}
\title{Beta}
\usage{
getBeta(x, id)

getBetaDissimilarity(x, ab)
}
\arguments{
\item{x}{(data.frame) Has to have columns Species, Year and Abundance or Biomass}

\item{id}{definition of id}

\item{ab}{character input for chosen currency - "A" = Abundance or "B" = Biomass}
}
\value{
getBeta returns a data.frame with three beta diversity dissimilarity
metrics

getBetaDissimilarity returns a long data.frame with results for three beta metrics
}
\description{
Beta

run the beta function
}
\examples{
\dontrun{
x <- data.frame(
  Year = rep(rep(2010:2015, each = 4), times = 4),
  Species = c(replicate(
   n = 8L,
   sample(letters, 24L, replace = FALSE))),
  Abundance = rpois(24 * 8, 10),
  ID = rep(LETTERS[1L:8L], each = 24)
  )
  res <- getBeta(x, "A")
}
\dontrun{
x <- data.frame(
  Year = rep(rep(2010:2015, each = 4), times = 4),
  Species = c(replicate(
   n = 8L,
   sample(letters, 24L, replace = FALSE))),
  Abundance = rpois(24 * 8, 10),
  rarefyID = rep(LETTERS[1L:8L], each = 24)
  )

res <- getBetaDissimilarity(x, "A")
}
}
\author{
Faye Moyes
}
