% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getAlpha}
\alias{getAlpha}
\alias{getAlphaMetrics}
\alias{getBeta}
\title{Alpha}
\usage{
getAlpha(x, getID)

getAlphaMetrics(x)

getBeta(x, id)
}
\arguments{
\item{x}{(data.frame) Has to have columns Species and Abundance}

\item{id}{definition of id}

\item{pivot}{should `x` be pivoted from long to wide? Default to FALSE.}
}
\value{
A data frame with results for S (species richness), N (numerical abundance),
maximum N per year per assemblage, Shannon, Exponential Shannon, Simpson,
Inverse Simpson, PIE (probability of intraspecific encounter) and McNaughton's Dominance
}
\description{
Alpha

run the alpha function

Beta
}
\examples{
x <- data.frame(
  Year = rep(rep(2010:2015, each = 4), times = 4),
  Species = unlist(lapply(
    X = 1L:8L,
    function(x) letters[sample(length(letters), 24L, replace = FALSE)])),
  Abundance = rpois(24 * 8, 10),
  ID = rep(LETTERS[1L:8L], each = 24)
)

res<-getAlphaMetrics(x)


res<-getAlphaMetrics(x)

res<-getBetaDissimilarity(x)
where x is a long form data frame
}
\author{
Faye Moyes
}
