% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getAlphav1}
\alias{getAlphav1}
\alias{getAlphav2}
\alias{getBetav2}
\title{Alpha v1}
\usage{
getAlphav1(x, id)

getAlphav2(x, id)

getBetav2(x, id)
}
\arguments{
\item{x}{(data.frame) First column has to be year}

\item{id}{definition of id}
}
\description{
Alpha v1

Alpha v2

Beta v2
}
\examples{
x <- data.frame(
  Year = rep(rep(2010:2015, each = 4), times = 4),
  Species = unlist(lapply(
    X = 1L:8L,
    function(x) letters[sample(length(letters), 24L, replace = FALSE)])),
  Abundance = rpois(24 * 8, 10),
  ID = rep(LETTERS[1L:8L], each = 24)
)
xd <- data.frame()

for (id in unique(x$ID)) {
  t11 <- base::subset(x, ID == id)
  t1 <- dplyr::select(t11, Year, Species, Abundance)
  xr <- getAlphav1(doPivot(t1), id)
  xd <- rbind(xd, xr)
}

xd <- data.frame()

for (id in unique(x$ID)) {
t11 <- base::subset(x, ID == id)
t1 <- dplyr::select(t11, Year, Species, Abundance)
xr <- getAlphav2(doPivot(t1), id)
xd <- rbind(xd, xr)
}
xd <- data.frame()

for (id in unique(x$ID)) {
  t11 <- subset(x, ID == id)
  t1 <- dplyr::select(t11, Year, Species, Abundance)
  xr <- getBetav2(doPivot(t1), id)
  xd <- rbind(xd, xr)
}

}
\author{
Faye Moyes
}
