% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsBioTIME.R
\name{plotSlopes}
\alias{plotSlopes}
\alias{themeBioTIME}
\title{Plot slopes BioTIME}
\usage{
plotSlopes(
  x,
  metric,
  cols,
  taxa = c("Amphibians & reptiles", "Birds", "Chromista", "Fish", "Fungi", "Mammals",
    "Plants"),
  method = c("metric", "taxa", "ind"),
  assemblageID,
  divType = c("alpha", "beta")
)

themeBioTIME(lp, fontSize, colx, coly)
}
\arguments{
\item{x}{Parameter description}

\item{metric}{If `divtype` is alpha, `metric` must be one of "S", "N", "maxN",
"Shannon", "Simpson", "invSimpson", "PIE", "DomMc","expShannon".
If `divtype` is beta, `metric` must be one of "JaccardDiss", "MorisitaHornDiss",
"BrayCurtisDiss"}

\item{cols}{Parameter description}

\item{taxa}{Parameter description}

\item{method}{Character can be one of "metric", "taxa", "ind"}

\item{assemblageID}{Parameter description}

\item{divType}{"alpha" or "beta"}

\item{lp}{description}

\item{fontSize}{description}

\item{colx}{description}

\item{coly}{description}
}
\value{
A plot
}
\description{
Plot slopes BioTIME

ggplot2 theme for BioTIME plots
}
\examples{
\donttest{
fig1 <- ggplot2::ggplot() +
          themeBioTIME("none", 12, "black", "grey90")
}
}
\keyword{internal}
