% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getAlpha}
\alias{getAlpha}
\title{Alpha}
\usage{
getAlpha(x, id)
}
\arguments{
\item{x}{(\code{data.frame}) First column has to be year and following columns
contain species abundances.}

\item{id}{(\code{character})One AssemblageID}
}
\value{
A data frame with results for S (species richness), N (numerical abundance),
maximum N per year per assemblage, Shannon, Exponential Shannon, Simpson,
Inverse Simpson, PIE (probability of intraspecific encounter) and
McNaughton's Dominance.
}
\description{
Alpha
}
\keyword{internal}
