% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getAlphaMetrics}
\alias{getAlphaMetrics}
\title{Run the alpha function}
\usage{
getAlphaMetrics(x, ab)
}
\arguments{
\item{x}{(data.frame) First column has to be year}

\item{ab}{character input for chosen currency - "A" = ABUNDANCE or "B" = BIOMASS}
}
\value{
getAlphaMetrics returns a data.frame with nine alpha diversity metrics
}
\description{
Run the alpha function
}
\examples{
\dontrun{
  x <- data.frame(
    resamp = 1L,
    YEAR = rep(rep(2010:2015, each = 4), times = 4),
    Species = c(replicate(n = 8L, sample(letters, 24L, replace = FALSE))),
    ABUNDANCE = rpois(24 * 8, 10),
    assemblageID = rep(LETTERS[1L:8L], each = 24)
  )
  res <- getAlphaMetrics(x, "A")
}
}
\author{
Faye Moyes
}
