% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getBeta}
\alias{getBeta}
\title{Beta}
\usage{
getBeta(x, id)
}
\arguments{
\item{x}{(\code{data.frame}) First column has to be year and following columns
contain species abundances.}

\item{id}{(\code{character}) One assemblageID}
}
\value{
getBeta returns a data.frame with three beta diversity dissimilarity
metrics
}
\description{
Beta
}
\keyword{internal}
