% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsBioTIME.R
\name{getBetaMetrics}
\alias{getBetaMetrics}
\title{run the beta function}
\usage{
getBetaMetrics(x, measure)
}
\arguments{
\item{x}{(data.frame) Has to have columns Species, YEAR, assemblageID,
STUDY_ID, cell and ABUNDANCE or BIOMASS}

\item{measure}{character input for chosen currency - "A" = ABUNDANCE or "B" = BIOMASS}
}
\value{
getBetaMetrics returns a long data.frame with results for three beta
  metrics.
}
\description{
run the beta function
}
\examples{
\dontrun{
x <- data.frame(
  resamp = 1L,
  YEAR = rep(rep(2010:2015, each = 4), times = 4),
  Species = c(replicate(
   n = 8L,
   sample(letters, 24L, replace = FALSE))),
  ABUNDANCE = rpois(24 * 8, 10),
  assemblageID = rep(LETTERS[1L:8L], each = 24)
  )

res <- getBetaMetrics(x, measure = "ABUNDANCE")
}
}
\author{
Faye Moyes
}
