% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopesBioTIME.R
\name{getLinearRegressions}
\alias{getLinearRegressions}
\title{Get Linear Regressions BioTIME}
\usage{
getLinearRegressions(x, divType)
}
\arguments{
\item{x}{output from getAlphaMetrics() or getBetaDissimilarity() functions}

\item{divType}{calculate regressions of alpha or beta results}
}
\value{
a dataframe with yearly diversity results merged with results of linear regressions (intercept, slope, p-value, significance)
}
\description{
Get Linear Regressions BioTIME
}
