% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopesBioTIME.R
\name{getLinearRegressions}
\alias{getLinearRegressions}
\title{Get Linear Regressions BioTIME
Fits linear regression models to \code{\link{getAlphaMetrics}} or \code{\link{getBetaMetrics}} outputs}
\usage{
getLinearRegressions(x, divType, pThreshold = 0.05)
}
\arguments{
\item{x}{(`data.frame`) BioTIME data table in the format of the output of  \code{\link{getAlphaMetrics}} or \code{\link{getBetaMetrics}}
functions}

\item{divType}{(`character`) string specifying the nature of the metrics in
the data; only `divType = "alpha"` or `divType = "beta"` is supported}

\item{pThreshold}{(`numeric`) P-value threshold for statistical significance}
}
\value{
Returns a single long `data.frame`, with the input diversity results concatenated with results of linear regressions (slope, pvalue, significance, intercept) for each year and `assemblageID`.
}
\description{
Get Linear Regressions BioTIME
Fits linear regression models to \code{\link{getAlphaMetrics}} or \code{\link{getBetaMetrics}} outputs
}
\details{
The function `getLinearRegressions` fits simple linear regression models (see \code{\link[stats]{lm}} for details) for a given output ('data') of either \code{\link{getAlphaMetrics}} or \code{\link{getBetaMetrics}} function.
`divType` needs to be specified in agreement with x.
The typical model has the form `metric ~ year`. Note that assemblages with less than 3 time points or/and single species time-series are removed.
}
\examples{
  x <- data.frame(
    resamp = 1L,
    YEAR = rep(rep(2010:2015, each = 4), times = 4),
    Species = c(replicate(n = 8L, sample(letters, 24L, replace = FALSE))),
    ABUNDANCE = rpois(24 * 8, 10),
    assemblageID = rep(LETTERS[1L:8L], each = 24)
  )
  alpham <- getAlphaMetrics(x, "ABUNDANCE")
  getLinearRegressions(x = alpham, divType = "alpha", pThreshold = 0.01)
  betam <- getBetaMetrics(x = x, "ABUNDANCE")
  getLinearRegressions(x = betam, divType = "beta")

}
