% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopesBioTIME.R
\name{getLinearRegressions}
\alias{getLinearRegressions}
\title{Get Linear Regressions BioTIME}
\usage{
getLinearRegressions(x, divType)
}
\arguments{
\item{x}{output from getAlphaMetrics() or getBetaDissimilarity() functions}

\item{divType}{calculate regressions of alpha or beta results}
}
\value{
a dataframe with yearly diversity results merged with results of
linear regressions (intercept, slope, p-value, significance)
}
\description{
Get Linear Regressions BioTIME
}
\examples{
  x <- data.frame(
    Year = rep(rep(2010:2015, each = 4), times = 4),
    Species = c(replicate(n = 8L, sample(letters, 24L, replace = FALSE))),
    Abundance = rpois(24 * 8, 10),
    rarefyID = rep(LETTERS[1L:8L], each = 24)
  )
  alpham <- getAlphaMetrics(x, "A")
  getLinearRegressions(x = alpham, divType = "alpha")
  betam <- getBetaDissimilarity(x = x, "A")
  getLinearRegressions(x = betam, divType = "beta")

}
