% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridBioTIME.R
\name{gridding}
\alias{gridding}
\title{gridding BioTIME}
\usage{
gridding(meta, btf)
}
\arguments{
\item{meta}{(data.frame) BioTIME metadata}

\item{btf}{(data.frame) BioTIME data}
}
\value{
A data.frame with studies split into 96km2 gridded cells (rarefyIDs)
}
\description{
gridding BioTIME
}
\examples{
\dontrun{
data("subBTquery.RData")
data("subBTmeta.RData")
gridding(subBTmeta, subBTquery)
}
}
