% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyBioTIME.R
\name{rarefysamples}
\alias{rarefysamples}
\title{rarefy BioTIME}
\usage{
rarefysamples(Year, SampleID, Species, Abundance, resamps)
}
\arguments{
\item{Year}{(character) string with latitude and longitude, one or more in a vector.}

\item{SampleID}{A definition of SampleID}

\item{Species}{A definition of Species}

\item{Abundance}{A definition of Abundance. All the same length?}

\item{resamps}{Number of repetition}
}
\value{
A data.frame with rarefied abundances
}
\description{
rarefy BioTIME
}
\examples{
\dontrun{
library(dplyr)

## select chosen rds file
dfr <- readRDS("savedFinalStage.rds")
## do any filtering required here
## for example:
## by length of study
rf <- subset(dfr, Abundance != 0)
yrf2 <- rf \%>\%
                        dplyr::group_by(rarefyID) \%>\%
                        dplyr::summarise(yr = dplyr::n_distinct(YEAR))
getYrs2 <- subset(yrf2, yr > 19)
rf2 <- subset(rf, rarefyID \%in\% getYrs2$rarefyID)
srf2 <- rf2 \%>\%
                        dplyr::group_by(rarefyID) \%>\%
                        dplyr::summarise(sp = dplyr::n_distinct(Species)) \%>\%
                        as.data.frame()
getSp2 <- subset(srf2, sp > 9)
## by taxa
## by realm
## by number of species
df <- dplyr::select(dfr, YEAR, Species, ABUNDANCE, rarefyID, SAMPLE_DESC)
stats::setNames(df, c("Year", "Species", "Abundance", "rarefyID", "Samp"))

TS <- df

TSrf <- list()
rfIDs <- unique(TS$rarefyID)
set.seed(123)
for (i in 1:length(rfIDs)) {
  data <- TS[TS$rarefyID == rfIDs[i],]
  TSrf[[i]] <- rarefysamples(Year = data$Year, SampleID = data$Samp,
                             Species = data$Species,
                             Abundance = data$Abundance, resamps = 1)
}
names(TSrf) <- rfIDs

rf <- do.call(rbind, TSrf)
rf <- data.frame(rf, rfID = rep(names(TSrf), times = unlist(lapply(TSrf, nrow))))
rf <- rf[!is.na(rf$Year),-1]
rownames(rf) <- NULL
rf1 <- rf \%>\%
  tidyr::separate(., rfID, into = c("STUDY_ID", "cell"),
                  sep = "_", remove = FALSE) \%>\%
  as.data.frame() \%>\%
  dplyr:::select(rf1, Year, Species, Abundance, rfID, STUDY_ID)
stats::setNames(rf1, c("Year", "Species", "Abundance", "rarefyID", "StudyID"))

## save file
#saveRDS(rf1, "savedRarefyv1.rds")
}
}
