% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyBioTIME.R
\name{rarefysamples}
\alias{rarefysamples}
\title{Rarefy BioTIME
Applies sample-based rarefaction to standardize the number of samples per year
   within a cell-level time-series.}
\usage{
rarefysamples(Year, SampleID, Species, currency, resamps)
}
\arguments{
\item{Year}{Integer representing year of record}

\item{SampleID}{Description of unique sampling event, this is used to resample}

\item{Species}{Highest resolution of taxonomic description available}

\item{currency}{Set to either (numeric) abundance or biomass}

\item{resamps}{Number of times the function randomly draws among the samples.}
}
\value{
Returns a single long form data frame containing the total currency
   of interest (sum) for each species in each year.
}
\description{
Rarefy BioTIME
Applies sample-based rarefaction to standardize the number of samples per year
   within a cell-level time-series.
}
\details{
Sample-based rarefaction prevents temporal variation in sampling effort
   from affecting diversity estimates [REF]. `rarefysamples()` is a function
   used to count the number of samples taken in each year, identify the
   minimum number of samples, and then use this minimum to randomly
   resample each year down to that number of samples, after which standard
   biodiversity metrics (e.g. \code{\link{getAlpha}}, \code{\link{getBeta}})
   can be calculated. Here sampling effort is defined as the number of
   samples (SampleID) taken in each year. `currency` is a numerical column
   with the variable (abundance or biomass) to be retained during the
   sample-based rarefaction, while `resamps` can be used to define multiple
   iterations (see examples). To apply directly to the output from the
   \code{\link{gridding}} function use \code{\link{runResampling}}.
}
\examples{
\dontrun{
  library(dplyr)
  library(BioTIMEr)
  df <- gridding(subBTmeta, subBTquery) \%>\%
    filter(assemblageID == "10_358678")
  rarefysamples(Year = df$YEAR, SampleID = df$SAMPLE_DESC,
    Species = df$Species, currency = df$ABUNDANCE,
    resamps = 3)
}

}
