% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyBioTIME.R
\name{rarefysamples}
\alias{rarefysamples}
\title{rarefy BioTIME}
\usage{
rarefysamples(Year, SampleID, Species, Abundance, resamps)
}
\arguments{
\item{Year}{Integer representing year of record}

\item{SampleID}{Description of unique sampling event, this is used to resample}

\item{Species}{Highest resolution of taxonomic description available}

\item{Abundance}{Numerical abundance}

\item{resamps}{Number of repetitions}
}
\value{
A data.frame with rarefied abundances
}
\description{
rarefy BioTIME
}
\examples{
\dontrun{
library(dplyr)

dfr <- # output from gridding process
df <- dplyr::select(dfr, YEAR, Species, ABUNDANCE, rarefyID, SAMPLE_DESC)
stats::setNames(df, c("Year", "Species", "Abundance", "rarefyID", "Samp"))
}



}
