% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyBioTIME.R
\name{rarefysamples}
\alias{rarefysamples}
\title{Rarefy BioTIME data
Applies sample-based rarefaction to standardise the number of samples per year
   within a cell-level time-series.}
\usage{
rarefysamples(df, measure, resamps)
}
\arguments{
\item{df}{`data.frame` to be resampled (in the format of the output of the
\code{\link{gridding}} function).}

\item{measure}{(character) Enter the names of the columns of interest:
"ABUNDANCE", "BIOMASS" or c("ABUNDANCE","BIOMASS"). The column cannot have 0s.}

\item{resamps}{(integer) Number of repetitions, 1 by default.}
}
\value{
Returns a single long form data frame containing the total currency
   of interest (sum) for each species in each year.
}
\description{
Rarefy BioTIME data
Applies sample-based rarefaction to standardise the number of samples per year
   within a cell-level time-series.
}
\details{
Sample-based rarefaction prevents temporal variation in sampling effort
   from affecting diversity estimates [REF]. `rarefysamples` is a function
   used to count the number of samples taken in each year, identify the
   minimum number of samples, and then use this minimum to randomly
   resample each year down to that number of samples, after which standard
   biodiversity metrics (e.g. \code{\link{getAlpha}}, \code{\link{getBeta}})
   can be calculated. Here sampling effort is defined as the number of
   samples (SampleID) taken in each year. `currency` is a numerical column
   with the variable (abundance or biomass) to be retained during the
   sample-based rarefaction, while `resamps` can be used to define multiple
   iterations (see examples). To apply directly to the output from the
   \code{\link{gridding}} function use \code{\link{resampling}}.
}
\examples{
\dontrun{
  library(dplyr)
  library(BioTIMEr)
  df <- gridding(subBTmeta, subBTquery) \%>\%
    filter(assemblageID == "10_358678")
  rarefysamples(Year = df$YEAR, SampleID = df$SAMPLE_DESC,
    Species = df$Species, currency = df$ABUNDANCE,
    resamps = 3)
}

}
\keyword{internal}
