% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyBioTIME.R
\name{rarefysamples}
\alias{rarefysamples}
\title{Rarefy BioTIME data
Applies sample-based rarefaction to standardise the number of samples per year
   within a cell-level time series.}
\usage{
rarefysamples(x, measure, summarise)
}
\arguments{
\item{x}{(\code{data.frame}) BioTIME gridded data to be resampled (in the format of
the output of the \code{\link{gridding}} function).}

\item{measure}{(\code{character}) currency to be retained during the sample-based
rarefaction. Can be either defined by a single column name or a vector of
two or more column names.}

\item{summarise}{(\code{logical}). \code{TRUE} by default. If \code{FALSE},
the function returns abundance and/or biomass summed at the}
}
\value{
Returns a single long form data frame containing the total currency
   of interest (sum) for each species in each year.
}
\description{
Rarefy BioTIME data
Applies sample-based rarefaction to standardise the number of samples per year
   within a cell-level time series.
}
\keyword{internal}
