% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyBioTIME.R
\name{rarefysamples}
\alias{rarefysamples}
\title{rarefy BioTIME}
\usage{
rarefysamples(Year, SampleID, Species, currency, resamps)
}
\arguments{
\item{Year}{Integer representing year of record}

\item{SampleID}{Description of unique sampling event, this is used to resample}

\item{Species}{Highest resolution of taxonomic description available}

\item{currency}{- set to either Numerical abundance or biomass}

\item{resamps}{Number of repetitions}
}
\value{
A data.frame with rarefied abundances/biomasses
}
\description{
rarefy BioTIME
}
\examples{
\dontrun{
library(dplyr)
library(BioTIMEr)
data("subBTmeta")
data("subBTquery")
df <- gridding(subBTmeta, subBTquery) \%>\%
  filter(rarefyID == "431_128327318")
rarefysamples(Year = df$YEAR, SampleID = df$SAMPLE_DESC,
  Species = df$Species, currency = df$ABUNDANCE,
  resamps = 3L)
}

}
