% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyBioTIME.R
\name{runResampling}
\alias{runResampling}
\title{runResampling BioTIME}
\usage{
runResampling(df, ab)
}
\arguments{
\item{df}{dataframe to be resampled (in the format of the output of the gridding() function}

\item{ab}{set to "A" for abundance and "B" for biomass}
}
\value{
dataframe containing rarefied studies
}
\description{
runResampling BioTIME
}
