% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyBioTIME.R
\name{runResampling}
\alias{runResampling}
\title{runResampling BioTIME}
\usage{
runResampling(df, ab, resamps = 1L)
}
\arguments{
\item{df}{dataframe to be resampled (in the format of the output of the
\code{\link{gridding}} function}

\item{ab}{set to "A" for abundance and "B" for biomass}

\item{resamps}{Not implemented at the moment.
Number of repetitions passed to \code{\link{rarefysamples}}. 1 by default.}
}
\value{
data.frame containing rarefied studies
}
\description{
runResampling BioTIME
}
\examples{
\dontrun{
  library(BioTIMEr)
  df<-gridding(subBTmeta, subBTquery)
  runResampling(df, ab = "A")
}

}
